requirements.txt
Flask>=2.2.0
Flask-Cors>=3.0.10
SQLAlchemy>=2.0.0
pymysql>=1.0.2
redis>=5.0.0
minio>=7.1.0
python-dotenv>=1.0.0
requests>=2.28.0
matplotlib>=3.5.0
numpy>=1.21.0
